payments_engine::account
Struct Account

pub struct Account {
    client_id: u16,
    available: Decimal,
    held: Decimal,
    locked: AccountState,
}

A financial account that tracks funds collected by a client.

The account maintains the following:

client_id: Unique identifier for the client.

available: Funds available for use.

held: Funds that are under dispute or reserved.

locked: State of the account (locked/unlocked).

Fields
client_id: u16available: Decimalheld: Decimallocked: AccountState
Implementations

impl Account

pub fn new(client_id: u16) -> Self

Creates a new account for the given client.

A new account is always created with an unique client_id and zero funds (both available and held) and unolocked state.

Arguments

client_id - Unique identifier for the client.

pub fn deposit(&mut self, amount: Amount) -> Result<(), Errors>

Deposits an amount into the account.

Increases the available funds by the specified amount if the account is unlocked. Returns an error if the account is locked or if the operation causes a funds overflow.

Arguments

amount - The amount to deposit.

pub fn withdrawal(&mut self, amount: Amount) -> Result<(), Errors>

Withdraws an amount from the account.

Decreases the available funds by the specified amount if sufficient funds exist and the account is unlocked. An error is returned if the account is locked, if there are insufficient funds, or if the operation causes an overflow.

Arguments

amount - The amount to withdraw.

pub fn dispute(&mut self, amount: Amount) -> Result<(), Errors>

Handles a dispute on an amount.

Blocks the disputed amount from available funds if enough available exist. Returns an error if there are insufficient funds to cover the dispute.

Arguments

amount - The amount to dispute.

pub fn resolve(&mut self, amount: Amount) -> Result<(), Errors>

Resolves a dispute by moving held funds back to available funds.

Increases the available funds and decreases the held funds by the disputed amount. Returns an error if the operation causes a funds overflow.

Arguments

amount - The amount to resolve.

pub fn chargeback(&mut self, amount: Amount) -> Result<(), Errors>

Processes a chargeback on a disputed amount.

Removes the disputed amount from held funds and locks the account. Returns an error if the operation causes a funds overflow.

Lock operation, as the effect of the chargeback, is done only when disputed amount doesn’t trigger overflow error to be reprorted.

Arguments

amount - The amount for the chargeback.

pub fn chargeback_withdrawal(&mut self, amount: Amount) -> Result<(), Errors>

Adjusts the account for a chargeback on a withdrawal operation.

Chargeback on withdrawal modifies only available funds, and doesn’t lock account once performed.

Arguments

amount - The amount to add back to the available funds.

fn lock(&mut self) -> Result<(), Errors>

Locks the account.