payments_engine::transaction_manager
Struct TransactionManager

pub struct TransactionManager {
    accounts: HashMap<u16, Account>,
    transactions: HashMap<u32, TransactionRecord>,
}

Provides business logic for this toy payments-engine. Controls overal flow over different types of transactions.

Internally contains only two collections: accounts and transactions. Provides implementation to properly handle different type of operations.

Fields
accounts: HashMap<u16, Account>transactions: HashMap<u32, TransactionRecord>
Implementations

impl TransactionManager

pub fn new() -> Self

Creates a new TransactionManager instance with empty account and transaction records.

fn get_account(&mut self, cliend_id: u16) -> &mut Account

Helper method to get a handler to an account (creating one if needed at first).

Arguments

cliend_id - The unique identifier of the client.

Returns

A mutable reference to the client’s Account.

pub fn parse_entry(&mut self, record: &Record) -> Result<(), Errors>

Parses a transaction record and updates the internal state accordingly.

The function handles different operation types:

Deposit/Withdrawal: Validates the record, ensures the transaction ID is unique, inserts the record, and updates the account balance.

Dispute: Flags a transaction as disputed and adjusts account funds if applicable.

Resolve: Reverses a dispute, moving held funds back to available funds.

Chargeback: Finalizes a dispute by removing held funds and locking the account.

There is a few assumptions:

transaction id must be unique, otherwise an error is reported

if the transaction provided to Deposit or Withdrawal operation is negative, then an error is reported

if there is no amount provided to Deposit or Withdrawal operation, then an error is reported

Resolve operation must be referenced to a dispute call, otherwise an error is reported

Chargeback operations run only over already disputed amounts (eg. there must be a precedeing Dispute operation)

Chargeback operation logic differs for referenced Deposit or Withdrawal operations

A single operation could be disputed only once, next are silently ignored

Arguments

record - A reference to the incoming transaction Record.

Errors

See assumptions.

pub fn accounts(&self) -> impl Iterator<Item = &Account>

Convience method, that returns iterator over accounts.